Renaming <elemento,listaenc>.hpp to Camel Case
These files already exists by default on VPL (Virtual Programming Lab),
and there they're called "Elemento.hpp" and "ListaEnc.hpp", so there
changes are required.
